[SP-188] refactor: PersonalityRepository 임포트 패키지 수정
