+++ src/main/java/com/cupid/jikting/team/service/TeamService.java
@@ -8,7 +8,7 @@ import com.cupid.jikting.team.dto.TeamRegisterResponse;
 import com.cupid.jikting.team.dto.TeamResponse;
 import com.cupid.jikting.team.dto.TeamUpdateRequest;
 import com.cupid.jikting.team.entity.Team;
-import com.cupid.jikting.team.repository.PersonalityRepository;
+import com.cupid.jikting.common.repository.PersonalityRepository;
 import com.cupid.jikting.team.repository.TeamRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
